requirements.txt
antlr4-python3-runtime==4.9.3
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
cssselect==1.1.0
Cython==0.29.32
dataclasses==0.8
Distance==0.1.3
emoji==1.2.0
fairseq==0.10.2
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.4.1
Flask==2.0.3
future==0.18.2
g2p-en==2.1.0
gensim==3.8.3
huggingface-hub==0.4.0
hydra-core==1.3.1
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
inflect==5.3.0
iniconfig==1.1.1
itsdangerous==2.0.1
jieba3k==0.35.1
Jinja2==3.0.3
joblib==1.1.1
JPype1==1.3.0
konlpy==0.6.0
kss==4.2.0
lxml==4.9.2
marisa-trie==0.7.8
MarkupSafe==2.0.1
mecab-python===0.996-ko-0.9.2
newspaper3k==0.2.8
nltk==3.6.7
numpy==1.19.5
omegaconf==2.3.0
packaging==21.3
pecab==1.0.8
Pillow==8.4.0
pluggy==1.0.0
pororo==0.4.2
portalocker==2.6.0
py==1.11.0
pyarrow==6.0.1
pybind11==2.9.2
pycparser==2.21
pyparsing==3.0.9
pytest==7.0.1
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.27.1
requests-file==1.5.1
sacrebleu==2.3.1
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.5.4
sentence-transformers==2.2.2
sentencepiece==0.1.97
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.3.2.post1
tabulate==0.8.10
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==3.1.2
tokenizers==0.12.1
tomli==1.2.3
torch==1.6.0
torchvision==0.7.0
tqdm==4.64.1
transformers==4.18.0
typing_extensions==4.1.1
urllib3==1.26.13
Werkzeug==2.0.3
wget==3.2
Whoosh==2.7.4
word2word==1.0.0
zipp==3.6.0
